refactor: 重构 main.ts 功能模块为独立文件并统一导出
