+++ src/modules/index.ts
@@ -1,0 +1,11 @@
+/**
+ * 功能模块统一入口
+ * 导出所有页面功能模块
+ */
+
+export { initDropdownMenus } from "./dropdown-menus";
+export { initBackToTop } from "./back-to-top";
+export { initMomentsTags } from "./moments-tags";
+export { initFloatingPagination } from "./floating-pagination";
+export { initActiveNavItem } from "./active-nav";
+export { initSearchShortcut } from "./search-shortcut";
